third_party/blink/web_tests/content_extraction/inline-wrap.html

<!DOCTYPE html>
<html>
<head>
<link href="./resources/consistent-rendering.css" rel="stylesheet" />
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./resources/content-extraction-test-helpers.js"></script>
<script src="../resources/ahem.js"></script>

<script>
function runTest() {
  finishTest(dumpContentNodeTree(document));
}
</script>
</head>
<body onload="runTest()">
 <body>
   <!--
   The [quick
   brown fox]
   [jumps over
   the lazy
   dog]
  -->
   <section style="width: 10ch;">
     The <b>quick brown fox</b> jumps over the lazy dog.
  </section>
</html>
